refactor: 重命名子目录名称
