+++ packages/kit/src/index.ts
@@ -1,4 +1,4 @@
 export * from './function/index'
+export * from './is/index'
 export * from './promise/index'
 export * from './tree/index'
-export * from './validate/index'
