+++ packages/kit/src/is/index.ts
