+++ packages/kit/src/is/is-circular/index.test.ts
