+++ packages/kit/src/is/is-circular/index.ts
