+++ packages/kit/src/is/is-digit/index.test.ts
