+++ packages/kit/src/is/is-digit/index.ts
