+++ packages/kit/src/is/is-email/index.test.ts
