+++ packages/kit/src/is/is-email/index.ts
