+++ packages/kit/src/is/is-firefox/index.test.ts
