+++ packages/kit/src/is/is-firefox/index.ts
