+++ packages/kit/src/is/is-safari/index.test.ts
