+++ packages/kit/src/is/is-safari/index.ts
